renaming things to make the edit version more clear
