+++ app/(app)/lish/[did]/[publication]/edit/[[...route]]/PublicationPagesEditNav.tsx
